fixup! refactor(calendar): rename base stylesheet
