+++ packages/components/src/Calendar/utils/setFocusInCalendar.ts
@@ -1,6 +1,6 @@
-import calendarStyles from "../Calendar.module.scss"
 import { CalendarRangeElement } from "../CalendarRange"
 import { CalendarSingleElement } from "../CalendarSingle"
+import calendarStyles from "../calendarClassNames.module.scss"
 import { isInvalidDate } from "./isInvalidDate"
 
 const isHTMLElement = (element: Element | undefined): element is HTMLElement =>
